+++ src/dijkstra.cpp
@@ -42,7 +42,7 @@ void Dijkstra::init(const DGraph *g) {
  * This assumes that the array d has been initialised with d[v] = INFINITE_DIST
  * for all vertices v != v0.
  */
-void Dijkstra::run(float *d, float *w, int *prev, unsigned int v0)
+void Dijkstra::run(double *d, double *w, int *prev, unsigned int v0)
 {
     /* indexes, counters, pointers */
     const DGraphEdge *edge;
@@ -80,7 +80,7 @@ void Dijkstra::run(float *d, float *w, int *prev, unsigned int v0)
             unsigned int et = edge->target;
 
             if (!s[et]) {
-                float wt = w [v] + edge->wt;
+                double wt = w [v] + edge->wt;
                 if (wt < w [et]) {
                     d [et] = d [v] + edge->dist;
                     w [et] = wt;
